+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 ## unreleased
 
 * [feat]: support `_bulk` api, fixes [#39][f39]
+* [feat]: can show, add and remove index aliases, fixes [#38][f38]
 
+[f38]: https://github.com/cars10/elasticvue/issues/38
 [f39]: https://github.com/cars10/elasticvue/issues/39